requirements.txt
Pillow
numpy
moviepy
Pygments
